% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_source}
\alias{data_source}
\title{Orignal source file as obtained from the ASPIS project cluster on
Google Drive}
\format{
A data frame with 790 rows and 12 variables:
\describe{
\item{Compound Name}{The generic name of the compound}
\item{IUPAC}{Standardized IUPAC chemical name}
\item{Inchi}{The inchi of the compound}
\item{SMILES}{The SMILES string to conceptualize the molecular structure}
\item{CAS number}{The number according the CAS registry}
\item{DTXSID}{EPA CompTox Dashboard compound id}
\item{Project ID (RISK-HUNT3R)}{Internal id for RISK-HUNT3R project}
\item{Project ID (ONTOX)}{Internal id for ONTOX project}
\item{Project ID (PrecisionTox)}{Internal id for PrecisionTox project}
\item{Used in RISK-HUNT3R}{Whether the compound is used in RISK-HUNT3R}
\item{Used in ONTOX}{Whether the compound is used in ONTOX}
\item{Used in PrecisionTox}{Whether the compound is used in PrecisionTox}

}
}
\source{
\url{Google Drive}
}
\usage{
data_source
}
\description{
A dataset containing information (ids) on 790 compounds, collected
from three Horizon 2020 projects (RISK-HUNT3R, ONTOX and PrecisionTox),
organized under an umbrella organization called ASPIS.
See \href{https://www.aspis-cluster.com/}{ASPIS} for more details.
}
\keyword{datasets}
