% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merged_source_cid_synonyms}
\alias{merged_source_cid_synonyms}
\title{Data set on compounds from the ASPIS cluster of projects,
enriched for PubChem CID en compound synonyms}
\format{
A data frame with 790 rows and 12 variables:
\describe{
\item{compound_name}{The generic name of the compound}
\item{iupac}{Standardized IUPAC chemical name}
\item{inchi}{The inchi of the compound}
\item{smiles}{The SMILES string to conceptualize the molecular structure}
\item{cas_number}{The number according the CAS registry}
\item{dtxsid}{EPA CompTox Dashboard compound id}
\item{project_id}{Internal id for projects}
\item{used_in_project}{Whether the compound is used in the project}
\item{cid}{Id of record in the
\href{https://pubchem.ncbi.nlm.nih.gov/}{PubChem Database}}
\item{synonyms}{Synonym of compound as retrieved from the
\href{https://pubchem.ncbi.nlm.nih.gov/}{PubChem Database}}
}
}
\source{
\url{Google Drive}
}
\usage{
merged_source_cid_synonyms
}
\description{
A data set containing information (ids) on 790 compounds, collected
from three Horizon 2020 projects (RISK-HUNT3R, ONTOX and PrecisionTox),
organized under an umbrella organization called ASPIS.
See \href{https://www.aspis-cluster.com/}{ASPIS} for more details on ASPIS.
}
\details{
This data set is an enriched version of dataset "data_source"
also available in \code{{raspis}}. The data was enriched with PubChem CID and
compound synonyms, obtained via the NCBI PUG-REST API).
Steps taken to do this enrichment are described in the package
Vignette. Run \code{browseVignettes(package = "raspis")} to see
the Vignette(s). There is also a tidy version of this data set available
as \code{raspis::merged_source_cid_synonyms_tidy}
}
\keyword{datasets}
